+++ CHANGELOG.md
@@ -12,7 +12,9 @@
 - BREAKING: Remove `subprojects` option; replace with `extraLinks` option –
   [#175](https://github.com/oddbird/sassdoc-theme-herman/issues/175)
 - BREAKING: Removed `@macro` annotation; use `@example njk` instead –
-  [#82](https://github.com/oddbird/sassdoc-theme-herman/issues/182)
+  [#182](https://github.com/oddbird/sassdoc-theme-herman/issues/182)
+- CHANGE: HTML assets are no longer minified –
+  [#186](https://github.com/oddbird/sassdoc-theme-herman/issues/186)
 
 
 ## 1.0.0-alpha.15: 2017-10-30